fix(runtime): handle clustered global options
